+++ test/blackbox-tests/test-cases/describe.t
@@ -215,6 +215,7 @@ Setup
   > EOF
   $ touch subdir/subfolder/subfolder_lib.ml
 
+
 Describe various things
 -----------------------
 
@@ -1304,6 +1305,7 @@ not stable across different setups.
         (cmt ())
         (cmti (_build/default/virtual/.virtual.objs/byte/virtual.cmti)))))
      (include_dirs (_build/default/virtual/.virtual.objs/byte)))))
+
 
 Test other formats
 ------------------